+++ CHANGES.md
@@ -1,6 +1,9 @@
 ## v0.0.6
   - :checkered_flag: **CHANGES**
     - Improved API Gateway `describe` output.
+    - Added [method response](http://docs.aws.amazon.com/apigateway/api-reference/resource/method-response/) support.  
+      - The [DefaultMethodResponses](https://godoc.org/github.com/mweagle/Sparta#DefaultMethodResponses) map is used if [Method.Responses](https://godoc.org/github.com/mweagle/Sparta#Method) is empty  (`len(Responses) <= 0`) at provision time.
+      - The default response map defines `201` for _POST_ methods, and `200` for all other methods. An API Gateway method may only support a single 2XX status code.
     - Added [integration response](http://docs.aws.amazon.com/apigateway/api-reference/resource/integration-response/) support for to support HTTP status codes defined in [status.go](https://golang.org/src/net/http/status.go).
       - The [DefaultIntegrationResponses](https://godoc.org/github.com/mweagle/Sparta#DefaultIntegrationResponses) map is used if [Integration.Responses](https://godoc.org/github.com/mweagle/Sparta#Integration) is empty  (`len(Responses) <= 0`) at provision time.
       - The mapping uses regular expressions based on the standard _golang_ HTTP StatusText values.